+++ DotnetCore/Sauce.Demo/Core.BestPractices.Web/DesktopWebPageObjects/ProductsPage.cs
@@ -1,7 +1,7 @@
 using System;
 using OpenQA.Selenium;
 
-namespace Core.BestPractices.Web.Pages
+namespace Core.BestPractices.Web.DesktopWebPageObjects
 {
     public class ProductsPage : BaseWebPage
     {
