+++ DotnetCore/Sauce.Demo/Core.BestPractices.Web/Tests/Desktop/DesktopTests.cs
@@ -1,4 +1,4 @@
-using Core.BestPractices.Web.Pages;
+using Core.BestPractices.Web.DesktopWebPageObjects;
 using FluentAssertions;
 using NUnit.Framework;
 using OpenQA.Selenium;
